resolved some PR related issues
